feat: variable interpolation
A variable like $VAR_NAME can be interpolated if the key VAR_NAME is
present in the snippet. The interpolated variable can be evaluated as
vim code, lua code or python code, or inserted verbatim.

To be evaluated as code, the content of the VAR_NAME key must begin
with '!vim', '!lua' or '!py'. Anything else is inserted verbatim.

Example:

  "date": {
      "body": [
          "$DATE"
      ],
      "description": "today's date",
      "prefix": [
          "date"
      ],
      "DATE": "!vim strftime('%D')"
  }